dev-requirements.txt
pytest-ckan
ckanext-scheming @ git+https://github.com/ckan/ckanext-scheming.git@release-3.0.0#egg=ckanext-scheming

djlint
isort
black